react and d3 integration
