+++ package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "dev": "next",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "fix-code": "prettier-eslint --write 'src/**/*.{js,jsx}' "
   },
   "author": "matt@numenta.org",
   "license": "AGPL-3.0",
@@ -21,5 +22,11 @@
     "react-syntax-highlighter": "^10.2.1",
     "simplehtm": "0.1.1"
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.16.0",
+    "eslint-plugin-react": "^7.13.0",
+    "prettier-eslint": "^8.8.2",
+    "prettier-eslint-cli": "^4.7.1"
+  }
 }
