+++ pages/components-showcase.jsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import Layout from '../components/Layout'
 import CodeSyntax from '../components/CodeSyntax'
 
@@ -30,14 +31,14 @@ module.exports = WeekendEncoder
 const highlightedLines = [5,6,8,10]
 
 export default function EncodingCategories() {
-  return (
-    <div>
-      <Layout>
-        <h2>Components Showcase</h2>
-        <i>This page demoes the available components for creating tutorials</i>
-        <h3>Code block</h3>
-        <CodeSyntax highlightedLines={highlightedLines}>{codeBlockContents}</CodeSyntax>
-      </Layout>
-    </div>
-  )
+	return (
+		<div>
+			<Layout>
+				<h2>Components Showcase</h2>
+				<i>This page demoes the available components for creating tutorials</i>
+				<h3>Code block</h3>
+				<CodeSyntax highlightedLines={highlightedLines}>{codeBlockContents}</CodeSyntax>
+			</Layout>
+		</div>
+	)
 }
