--- pages/encoding-categories.js
@@ -1,11 +1,0 @@
-import Layout from '../components/Layout'
-
-export default function EncodingCategories() {
-  return (
-    <div>
-      <Layout>
-        <h2>Encoding Categories</h2>
-      </Layout>
-    </div>
-  )
-}
