+++ pages/encoding-categories.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Layout from '../components/Layout'
+
+export default function EncodingCategories() {
+	return (
+		<div>
+			<Layout>
+				<h2>Encoding Categories</h2>
+			</Layout>
+		</div>
+	)
+}
