--- pages/encoding-time.js
@@ -1,11 +1,0 @@
-import Layout from '../components/Layout'
-
-export default function EncodingTime() {
-  return (
-    <div>
-      <Layout>
-        <h2>Encoding Time</h2>
-      </Layout>
-    </div>
-  )
-}
