+++ pages/encoding-time.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Layout from '../components/Layout'
+
+export default function EncodingTime() {
+	return (
+		<div>
+			<Layout>
+				<h2>Encoding Time</h2>
+			</Layout>
+		</div>
+	)
+}
