+++ pages/index.jsx
@@ -1,0 +1,22 @@
+import React from 'react'
+import Layout from '../components/Layout'
+
+const Index = () => (
+	<div>
+		<Layout>
+			<h1>Building HTM Systems</h1>
+			<p>
+        This website is under active development on <a href="https://www.twitch.tv/rhyolight_">Twitch</a>.
+			</p>
+			<ul>
+				<li><a href="https://www.twitch.tv/rhyolight_/events">livestream schedule</a></li>
+				<li><a href="https://github.com/htm-community/building-htm-systems">github</a></li>
+				<li><a href="https://github.com/htm-community/simplehtm">simplehtm</a></li>
+				<li><a href="https://trello.com/b/iOG0iEnT/building-htm-systems ">trello</a></li>
+				<li><a href="https://travis-ci.org/htm-community/building-htm-systems">travis</a></li>
+			</ul>
+		</Layout>
+	</div>
+)
+
+export default Index
